pyproject.toml
[tool.poetry]
name = "slack-timezoner"
version = "0.1.0"
description = ""
authors = ["Chris Adams <chris@productscience.co.uk>"]

[tool.poetry.dependencies]
python = ">=3.6"
django = "^3.0.8"
python-dotenv = "^0.14.0"
slackclient = "^2.7.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipdb = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
